Altereção de Pastas, ajustes de dtos e Controller Raca

+++ src/main/java/rpg/api/models/DadosListagemPersonagem.java
@@ -5,7 +5,7 @@ public record DadosListagemPersonagem(
         Long id,
         String nome,
         int idclasse,
-        int idraca){
+        Long idraca){
 
     public DadosListagemPersonagem(Personagem pp) {
         this( pp.getId() ,pp.getNome(), pp.getIdClasse(), pp.getIdRaca()
